LPS-130155 Remove unused code and reorganize leftovers
