+++ modules/apps/frontend-taglib/frontend-taglib-clay/src/main/java/com/liferay/frontend/taglib/clay/internal/servlet/ServletContextUtil.java
@@ -14,18 +14,7 @@
 
 package com.liferay.frontend.taglib.clay.internal.servlet;
 
-import com.liferay.frontend.taglib.clay.data.set.ClayDataSetDisplayViewSerializer;
-import com.liferay.frontend.taglib.clay.data.set.filter.ClayDataSetFilterSerializer;
-import com.liferay.frontend.taglib.clay.servlet.taglib.DataSetDisplayTag;
-import com.liferay.petra.string.StringBundler;
-import com.liferay.petra.string.StringPool;
-import com.liferay.portal.kernel.theme.PortletDisplay;
-import com.liferay.portal.kernel.theme.ThemeDisplay;
-import com.liferay.portal.kernel.util.Portal;
-import com.liferay.portal.kernel.util.WebKeys;
-
 import javax.servlet.ServletContext;
-import javax.servlet.http.HttpServletRequest;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -36,68 +25,12 @@ import org.osgi.service.component.annotations.Reference;
 @Component(service = {})
 public class ServletContextUtil {
 
-	public static String getClayDataSetDisplaySettingsNamespace(
-		HttpServletRequest httpServletRequest, String id) {
-
-		StringBundler sb = new StringBundler(7);
-
-		sb.append(DataSetDisplayTag.class.getName());
-		sb.append(StringPool.POUND);
-
-		ThemeDisplay themeDisplay =
-			(ThemeDisplay)httpServletRequest.getAttribute(
-				WebKeys.THEME_DISPLAY);
-
-		PortletDisplay portletDisplay = themeDisplay.getPortletDisplay();
-
-		String portletNamespace = _portal.getPortletNamespace(
-			portletDisplay.getId());
-
-		sb.append(portletNamespace);
-
-		sb.append(StringPool.POUND);
-		sb.append(themeDisplay.getPlid());
-		sb.append(StringPool.POUND);
-		sb.append(id);
-
-		return sb.toString();
-	}
-
-	public static ClayDataSetDisplayViewSerializer
-		getClayDataSetDisplayViewSerializer() {
-
-		return _clayDataSetDisplayViewSerializer;
-	}
-
-	public static ClayDataSetFilterSerializer getClayDataSetFilterSerializer() {
-		return _clayDataSetFilterSerializer;
-	}
-
 	public static String getContextPath() {
 		return _servletContext.getContextPath();
 	}
 
 	public static ServletContext getServletContext() {
 		return _servletContext;
-	}
-
-	@Reference(unbind = "-")
-	protected void setClayDataSetDisplayViewSerializer(
-		ClayDataSetDisplayViewSerializer clayDataSetDisplayViewSerializer) {
-
-		_clayDataSetDisplayViewSerializer = clayDataSetDisplayViewSerializer;
-	}
-
-	@Reference(unbind = "-")
-	protected void setClayDataSetFilterSerializer(
-		ClayDataSetFilterSerializer clayDataSetFilterSerializer) {
-
-		_clayDataSetFilterSerializer = clayDataSetFilterSerializer;
-	}
-
-	@Reference(unbind = "-")
-	protected void setPortal(Portal portal) {
-		_portal = portal;
 	}
 
 	@Reference(
@@ -108,10 +41,6 @@ public class ServletContextUtil {
 		_servletContext = servletContext;
 	}
 
-	private static ClayDataSetDisplayViewSerializer
-		_clayDataSetDisplayViewSerializer;
-	private static ClayDataSetFilterSerializer _clayDataSetFilterSerializer;
-	private static Portal _portal;
 	private static ServletContext _servletContext;
 
 }
